Don't export / duplicate to ROMFS in PKG mode (more of a issue to be solved in meds)
