+++ open_samus_returns_rando/specific_patches/door_patches.py
@@ -4,7 +4,7 @@ import copy
 from construct import Container, ListContainer
 from mercury_engine_data_structures.formats import Bmsad
 
-from open_samus_returns_rando.constants import ALL_AREAS
+from open_samus_returns_rando.constants import lua_pkgs
 from open_samus_returns_rando.files import files_path
 from open_samus_returns_rando.patcher_editor import PatcherEditor
 
@@ -97,9 +97,10 @@ def _patch_beam_bmsads(editor: PatcherEditor):
             cr_bmsad.components["COLLISION"].functions[1].params["Param10"].value = 300.0
         editor.replace_asset(creature_bmsad_file, cr_bmsad)
 
-    editor.add_new_asset("actors/props/doors/scripts/doors.lc", files_path().joinpath("doors.lua").read_bytes(), [])
-    for area_name in ALL_AREAS:
-        editor.ensure_present_in_scenario(area_name, "actors/props/doors/scripts/doors.lc")
+    editor.add_new_asset(
+        "actors/props/doors/scripts/doors.lc", files_path().joinpath("doors.lua").read_bytes(),
+        lua_pkgs(editor.get_all_level_pkgs())
+    )
 
 
 def _patch_beam_covers(editor: PatcherEditor):
